+++ src/DDTrace/Util/TryCatchFinally.php
@@ -46,6 +46,45 @@ class TryCatchFinally
     }
 
     /**
+     * PHP 5.4 compatible try-catch-finally to execute instance methods.
+     *
+     * @param Scope $scope
+     * @param mixed $instance
+     * @param string $method
+     * @param array $args
+     * @param \Closure $afterResult
+     * @return mixed|null
+     * @throws \Exception
+     */
+    public static function executeAnyMethod(Scope $scope, $instance, $method, array $args = [], $afterResult = null)
+    {
+        $thrown = null;
+        $result = null;
+        /** @var Span $span */
+        $span = $scope->getSpan();
+
+        try {
+            $result = call_user_func(\Closure::bind(function () use ($instance, $method, $args) {
+                return call_user_func_array([$instance, $method], $args);
+            }, null, $instance));
+
+            if ($afterResult) {
+                $afterResult($result, $span, $scope);
+            }
+        } catch (\Exception $ex) {
+            $thrown = $ex;
+            $span->setError($ex);
+        }
+
+        $scope->close();
+        if ($thrown) {
+            throw $thrown;
+        }
+
+        return $result;
+    }
+
+    /**
      * PHP 5.4 compatible try-catch-finally to execute functions.
      *
      * @param Scope $scope